requirements.txt
joblib
better-profanity
requests
json
python-dotenv
streamlit
openai
